% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{dataDTWBI}
\alias{dataDTWBI}
\title{Six univariate signals as example for DTWBI package}
\format{A data frame with six variables: \code{query}, \code{ref1},
  \code{ref2}, \code{ref3}, \code{ref4} and \code{ref5}.}
\usage{
dataDTWBI
}
\description{
Query and ref1 are two dephased sigmoidal signals.
Ref2 presents a linear decrease.
Ref3 and ref4 are constant signals of value 3 and 0 respectively.
Ref5 is similar to the query with small noise added.
}
\keyword{datasets}
